# Replace internal utility functions