@milmazz 2015-09-02
With `Integer.to_char_list(head, 16)` the result is a little bit slower:

Before:
```
uuid3 dns              500000   3.91 µs/op
binary_to_string!      500000   4.11 µs/op
uuid5 dns              500000   4.55 µs/op
uuid4                  500000   4.97 µs/op
string_to_binary!      500000   5.76 µs/op
info!                  500000   7.00 µs/op
uuid1                   10000   137.68 µs/op
```

After:
```
## UUIDBench
binary_to_string!      500000   4.32 µs/op
uuid3 dns              500000   4.57 µs/op
uuid5 dns              500000   5.04 µs/op
string_to_binary!      500000   5.22 µs/op
uuid4                  500000   5.34 µs/op
info!                  500000   5.82 µs/op
uuid1                   10000   146.86 µs/op
```

So, this appears to be a fight of clarity against performance, please, feel free to close this PR.