@zyro 2015-09-07
@milmazz At this point I'll take the extra speed over extra clarity. I expect this code to quietly sit in the background and generate IDs, with not much need to open it up.