@milmazz 2015-09-07
@zyro It would be nice if you put a list of goals for this project (e.g. performance over clarity) in some section on the README file or something like the CONTRIBUTING file, also some advice to the developers (e.g. Before to create a PR run `mix bench` to check the performance or something like that) would be awesome.